+++ examples/resources/meraki_switch_routing_interfaces/import.sh
@@ -1,0 +1,1 @@
+terraform import meraki_switch_routing_interfaces.example "<organization_id>,<serial>"