+++ internal/provider/resource_meraki_switch_routing_interfaces.go
@@ -26,11 +26,13 @@ import (
 	"strings"
 
 	"github.com/CiscoDevNet/terraform-provider-meraki/internal/provider/helpers"
+	"github.com/hashicorp/terraform-plugin-framework-validators/stringvalidator"
 	"github.com/hashicorp/terraform-plugin-framework/path"
 	"github.com/hashicorp/terraform-plugin-framework/resource"
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema"
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema/planmodifier"
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema/stringplanmodifier"
+	"github.com/hashicorp/terraform-plugin-framework/schema/validator"
 	"github.com/hashicorp/terraform-plugin-framework/types"
 	"github.com/hashicorp/terraform-plugin-log/tflog"
 	"github.com/netascode/go-meraki"
@@ -42,26 +44,26 @@ import (
 
 // Ensure provider defined types fully satisfy framework interfaces
 var (
-	_ resource.Resource                = &OrganizationPolicyObjectGroupsResource{}
-	_ resource.ResourceWithImportState = &OrganizationPolicyObjectGroupsResource{}
+	_ resource.Resource                = &SwitchRoutingInterfacesResource{}
+	_ resource.ResourceWithImportState = &SwitchRoutingInterfacesResource{}
 )
 
-func NewOrganizationPolicyObjectGroupsResource() resource.Resource {
-	return &OrganizationPolicyObjectGroupsResource{}
+func NewSwitchRoutingInterfacesResource() resource.Resource {
+	return &SwitchRoutingInterfacesResource{}
 }
 
-type OrganizationPolicyObjectGroupsResource struct {
+type SwitchRoutingInterfacesResource struct {
 	client *meraki.Client
 }
 
-func (r *OrganizationPolicyObjectGroupsResource) Metadata(ctx context.Context, req resource.MetadataRequest, resp *resource.MetadataResponse) {
-	resp.TypeName = req.ProviderTypeName + "_organization_policy_object_groups"
+func (r *SwitchRoutingInterfacesResource) Metadata(ctx context.Context, req resource.MetadataRequest, resp *resource.MetadataResponse) {
+	resp.TypeName = req.ProviderTypeName + "_switch_routing_interfaces"
 }
 
-func (r *OrganizationPolicyObjectGroupsResource) Schema(ctx context.Context, req resource.SchemaRequest, resp *resource.SchemaResponse) {
+func (r *SwitchRoutingInterfacesResource) Schema(ctx context.Context, req resource.SchemaRequest, resp *resource.SchemaResponse) {
 	resp.Schema = schema.Schema{
 		// This description is used by the documentation generator and the language server.
-		MarkdownDescription: helpers.NewAttributeDescription("This resource can manage the `Organization Policy Object Group` configuration.").String,
+		MarkdownDescription: helpers.NewAttributeDescription("This resource can manage the `Switch Routing Interface` configuration.").String,
 
 		Attributes: map[string]schema.Attribute{
 			"id": schema.StringAttribute{
@@ -72,7 +74,11 @@ func (r *OrganizationPolicyObjectGroupsResource) Schema(ctx context.Context, req
 				},
 			},
 			"organization_id": schema.StringAttribute{
-				MarkdownDescription: helpers.NewAttributeDescription("Organization ID").String,
+				MarkdownDescription: "The organization ID",
+				Required:            true,
+			},
+			"serial": schema.StringAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Switch serial").String,
 				Required:            true,
 			},
 			"items": schema.ListNestedAttribute{
@@ -87,17 +93,59 @@ func (r *OrganizationPolicyObjectGroupsResource) Schema(ctx context.Context, req
 								stringplanmodifier.UseStateForUnknown(),
 							},
 						},
-						"category": schema.StringAttribute{
-							MarkdownDescription: helpers.NewAttributeDescription("Category of a policy object group (one of: NetworkObjectGroup, GeoLocationGroup, PortObjectGroup, ApplicationGroup)").String,
+						"default_gateway": schema.StringAttribute{
+							MarkdownDescription: helpers.NewAttributeDescription("The next hop for any traffic that isn`t going to a directly connected subnet or over a static route. This IP address must exist in a subnet with a routed interface. Required if this is the first IPv4 interface.").String,
 							Optional:            true,
 						},
+						"interface_ip": schema.StringAttribute{
+							MarkdownDescription: helpers.NewAttributeDescription("The IP address this switch will use for layer 3 routing on this VLAN or subnet. This cannot be the same as the switch`s management IP.").String,
+							Optional:            true,
+						},
+						"multicast_routing": schema.StringAttribute{
+							MarkdownDescription: helpers.NewAttributeDescription("Enable multicast support if, multicast routing between VLANs is required. Options are: `disabled`, `enabled` or `IGMP snooping querier`. Default is `disabled`.").AddStringEnumDescription("IGMP snooping querier", "disabled", "enabled").String,
+							Optional:            true,
+							Validators: []validator.String{
+								stringvalidator.OneOf("IGMP snooping querier", "disabled", "enabled"),
+							},
+						},
 						"name": schema.StringAttribute{
-							MarkdownDescription: helpers.NewAttributeDescription("A name for the group of network addresses, unique within the organization (alphanumeric, space, dash, or underscore characters only)").String,
+							MarkdownDescription: helpers.NewAttributeDescription("A friendly name or description for the interface or VLAN.").String,
 							Required:            true,
 						},
-						"object_ids": schema.SetAttribute{
-							MarkdownDescription: helpers.NewAttributeDescription("A list of Policy Object ID`s that this NetworkObjectGroup should be associated to (note: these ID`s will replace the existing associated Policy Objects)").String,
-							ElementType:         types.Int64Type,
+						"subnet": schema.StringAttribute{
+							MarkdownDescription: helpers.NewAttributeDescription("The network that this routed interface is on, in CIDR notation (ex. 10.1.1.0/24).").String,
+							Optional:            true,
+						},
+						"vlan_id": schema.Int64Attribute{
+							MarkdownDescription: helpers.NewAttributeDescription("The VLAN this routed interface is on. VLAN must be between 1 and 4094.").String,
+							Optional:            true,
+						},
+						"ipv6_address": schema.StringAttribute{
+							MarkdownDescription: helpers.NewAttributeDescription("The IPv6 address of the interface. Required if assignmentMode is `static`. Must not be included if assignmentMode is `eui-64`.").String,
+							Optional:            true,
+						},
+						"ipv6_assignment_mode": schema.StringAttribute{
+							MarkdownDescription: helpers.NewAttributeDescription("The IPv6 assignment mode for the interface. Can be either `eui-64` or `static`.").String,
+							Optional:            true,
+						},
+						"ipv6_gateway": schema.StringAttribute{
+							MarkdownDescription: helpers.NewAttributeDescription("The IPv6 default gateway of the interface. Required if prefix is defined and this is the first interface with IPv6 configured for the switch.").String,
+							Optional:            true,
+						},
+						"ipv6_prefix": schema.StringAttribute{
+							MarkdownDescription: helpers.NewAttributeDescription("The IPv6 prefix of the interface. Required if IPv6 object is included.").String,
+							Optional:            true,
+						},
+						"ospf_settings_area": schema.StringAttribute{
+							MarkdownDescription: helpers.NewAttributeDescription("The OSPF area to which this interface should belong. Can be either `ospfDisabled` or the identifier of an existing OSPF area. Defaults to `ospfDisabled`.").String,
+							Optional:            true,
+						},
+						"ospf_settings_cost": schema.Int64Attribute{
+							MarkdownDescription: helpers.NewAttributeDescription("The path cost for this interface. Defaults to 1, but can be increased up to 65535 to give lower priority.").String,
+							Optional:            true,
+						},
+						"ospf_settings_is_passive_enabled": schema.BoolAttribute{
+							MarkdownDescription: helpers.NewAttributeDescription("When enabled, OSPF will not run on the interface, but the subnet will still be advertised.").String,
 							Optional:            true,
 						},
 					},
@@ -107,7 +155,7 @@ func (r *OrganizationPolicyObjectGroupsResource) Schema(ctx context.Context, req
 	}
 }
 
-func (r *OrganizationPolicyObjectGroupsResource) Configure(_ context.Context, req resource.ConfigureRequest, _ *resource.ConfigureResponse) {
+func (r *SwitchRoutingInterfacesResource) Configure(_ context.Context, req resource.ConfigureRequest, _ *resource.ConfigureResponse) {
 	if req.ProviderData == nil {
 		return
 	}
@@ -119,8 +167,8 @@ func (r *OrganizationPolicyObjectGroupsResource) Configure(_ context.Context, re
 
 // Section below is generated&owned by "gen/generator.go". //template:begin create
 
-func (r *OrganizationPolicyObjectGroupsResource) Create(ctx context.Context, req resource.CreateRequest, resp *resource.CreateResponse) {
-	var plan ResourceOrganizationPolicyObjectGroups
+func (r *SwitchRoutingInterfacesResource) Create(ctx context.Context, req resource.CreateRequest, resp *resource.CreateResponse) {
+	var plan ResourceSwitchRoutingInterfaces
 
 	// Read plan
 	diags := req.Plan.Get(ctx, &plan)
@@ -136,7 +184,7 @@ func (r *OrganizationPolicyObjectGroupsResource) Create(ctx context.Context, req
 		actions[i] = meraki.ActionModel{
 			Operation: "create",
 			Resource:  plan.getPath(),
-			Body:      item.toBody(ctx, ResourceOrganizationPolicyObjectGroupsItems{}),
+			Body:      item.toBody(ctx, ResourceSwitchRoutingInterfacesItems{}),
 		}
 	}
 	res, err := r.client.Batch(plan.OrganizationId.ValueString(), actions)
@@ -161,8 +209,8 @@ func (r *OrganizationPolicyObjectGroupsResource) Create(ctx context.Context, req
 
 // Section below is generated&owned by "gen/generator.go". //template:begin read
 
-func (r *OrganizationPolicyObjectGroupsResource) Read(ctx context.Context, req resource.ReadRequest, resp *resource.ReadResponse) {
-	var state ResourceOrganizationPolicyObjectGroups
+func (r *SwitchRoutingInterfacesResource) Read(ctx context.Context, req resource.ReadRequest, resp *resource.ReadResponse) {
+	var state ResourceSwitchRoutingInterfaces
 
 	// Read state
 	diags := req.State.Get(ctx, &state)
@@ -205,8 +253,8 @@ func (r *OrganizationPolicyObjectGroupsResource) Read(ctx context.Context, req r
 
 // Section below is generated&owned by "gen/generator.go". //template:begin update
 
-func (r *OrganizationPolicyObjectGroupsResource) Update(ctx context.Context, req resource.UpdateRequest, resp *resource.UpdateResponse) {
-	var plan, state ResourceOrganizationPolicyObjectGroups
+func (r *SwitchRoutingInterfacesResource) Update(ctx context.Context, req resource.UpdateRequest, resp *resource.UpdateResponse) {
+	var plan, state ResourceSwitchRoutingInterfaces
 
 	// Read plan
 	diags := req.Plan.Get(ctx, &plan)
@@ -265,7 +313,7 @@ func (r *OrganizationPolicyObjectGroupsResource) Update(ctx context.Context, req
 			actions = append(actions, meraki.ActionModel{
 				Operation: "create",
 				Resource:  plan.getPath(),
-				Body:      item.toBody(ctx, ResourceOrganizationPolicyObjectGroupsItems{}),
+				Body:      item.toBody(ctx, ResourceSwitchRoutingInterfacesItems{}),
 			})
 		}
 	}
@@ -286,8 +334,8 @@ func (r *OrganizationPolicyObjectGroupsResource) Update(ctx context.Context, req
 
 // Section below is generated&owned by "gen/generator.go". //template:begin delete
 
-func (r *OrganizationPolicyObjectGroupsResource) Delete(ctx context.Context, req resource.DeleteRequest, resp *resource.DeleteResponse) {
-	var state ResourceOrganizationPolicyObjectGroups
+func (r *SwitchRoutingInterfacesResource) Delete(ctx context.Context, req resource.DeleteRequest, resp *resource.DeleteResponse) {
+	var state ResourceSwitchRoutingInterfaces
 
 	// Read state
 	diags := req.State.Get(ctx, &state)
@@ -319,17 +367,18 @@ func (r *OrganizationPolicyObjectGroupsResource) Delete(ctx context.Context, req
 // End of section. //template:end delete
 
 // Section below is generated&owned by "gen/generator.go". //template:begin import
-func (r *OrganizationPolicyObjectGroupsResource) ImportState(ctx context.Context, req resource.ImportStateRequest, resp *resource.ImportStateResponse) {
+func (r *SwitchRoutingInterfacesResource) ImportState(ctx context.Context, req resource.ImportStateRequest, resp *resource.ImportStateResponse) {
 	idParts := strings.Split(req.ID, ",")
 
-	if len(idParts) != 1 || idParts[0] == "" {
+	if len(idParts) != 2 || idParts[0] == "" || idParts[1] == "" {
 		resp.Diagnostics.AddError(
 			"Unexpected Import Identifier",
-			fmt.Sprintf("Expected import identifier with format: <organization_id>. Got: %q", req.ID),
+			fmt.Sprintf("Expected import identifier with format: <organization_id>,<serial>. Got: %q", req.ID),
 		)
 		return
 	}
 	resp.Diagnostics.Append(resp.State.SetAttribute(ctx, path.Root("organization_id"), idParts[0])...)
+	resp.Diagnostics.Append(resp.State.SetAttribute(ctx, path.Root("serial"), idParts[1])...)
 
 	helpers.SetFlagImporting(ctx, true, resp.Private, &resp.Diagnostics)
 }